requirements.txt
flask
gspread
google-auth
reportlab
gunicorn